+++ sunray/Stats.cpp
@@ -62,7 +62,7 @@ void calcStats(){
           if (lastSolution == SOL_INVALID) statMowInvalidRecoveries++;
           lastSolution = gps.solution;
         } 
-        statMowMaxDgpsAge = max(statMowMaxDgpsAge, (millis() - gps.dgpsAge)/1000.0);        
+        statMowMaxDgpsAge = max(statMowMaxDgpsAge, (float)((millis() - gps.dgpsAge)/1000.0));        
         break;
       case OP_CHARGE:
         statChargeDuration++;
